requirements.txt
asgiref==3.8.1
Django==5.1
Faker==33.1.0
pillow==11.3.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
